[LP-464] refactor: gallery detail filter page 비즈니스 로직을 service와 hooks로 분리
- 이미지 처리 유틸리티 함수들을 image-processing.service.ts로 분리
- 필터 관련 로직을 image-filter.service.ts로 분리
- 이미지 상태 관리를 위한 image-state.hook.ts 추가
- 필터 작업을 위한 filter-operations.hook.ts 추가
- GalleryDetailFilterPage 컴포넌트 코드 170줄에서 73줄로 단순화
- 재사용성과 유지보수성 향상
