+++ app/pages/StoryGallerySelector/GalleryDetailFilterPage.tsx
@@ -1,31 +1,15 @@
-import {useCallback, useEffect, useState} from 'react';
-import {
-  Dimensions,
-  Platform,
-  ScrollView,
-  TouchableOpacity,
-  View,
-  Image,
-} from 'react-native';
+import React, {useEffect, useCallback, useMemo} from 'react';
+import {ScrollView, TouchableOpacity, View} from 'react-native';
 import {ActivityIndicator} from 'react-native-paper';
 import {useNavigation} from '@react-navigation/native';
 import {BasicNavigationProps} from '../../navigation/types.tsx';
-import {useRecoilState} from 'recoil';
-import {selectedGalleryIndexState} from '../../recoils/photos.recoil.ts';
-import {editedGalleryItemsState} from '../../recoils/gallery-write.recoil.ts';
 import Slider from '@react-native-community/slider';
-import RNFS, {writeFile} from 'react-native-fs';
-import {encode as encodeBase64} from 'base64-arraybuffer';
-import PhotoManipulator from 'react-native-photo-manipulator';
-import {PhotoIdentifier} from '@react-native-camera-roll/camera-roll';
 import {
   Canvas,
-  SkImage,
   Image as SkiaImage,
   ColorMatrix,
   Blur,
   useCanvasRef,
-  Skia,
 } from '@shopify/react-native-skia';
 import {LoadingContainer} from '../../components/loadding/LoadingContainer.tsx';
 import {ScreenContainer} from '../../components/styled/container/ScreenContainer.tsx';
@@ -33,213 +17,61 @@ import {ContentContainer} from '../../components/styled/container/ContentContain
 import {TopBar} from '../../components/styled/components/TopBar.tsx';
 import WritingHeaderRight from '../../components/header/WritingHeaderRight.tsx';
 import {Title} from '../../components/styled/components/Text.tsx';
-import {CustomAlert} from '../../components/alert/CustomAlert.tsx';
 import {Color} from '../../constants/color.constant.ts';
 import {
   FILTER_LABELS,
   FILTER_EFFECTS,
-  FILTER_SETTINGS,
   FilterType,
 } from '../../constants/filter.constant.ts';
-
-const {width: screenWidth} = Dimensions.get('window');
-const displaySize = screenWidth;
-
-//TODO: 리팩터링 필요
-const getImageSizeAsync = (
-  uri: string,
-): Promise<{width: number; height: number}> =>
-  new Promise((resolve, reject) => {
-    Image.getSize(
-      uri,
-      (width, height) => resolve({width, height}),
-      error => reject(error),
-    );
-  });
-
-async function copyContentUriToFile(
-  selectedImage: PhotoIdentifier,
-): Promise<string> {
-  const uri = selectedImage.node.image.uri;
-  if (Platform.OS === 'android' && uri.startsWith('content://')) {
-    const dest = `${RNFS.TemporaryDirectoryPath}/${Date.now()}.jpg`;
-    await RNFS.copyFile(uri, dest);
-    return `file://${dest}`;
-  }
-  if (Platform.OS === 'ios' && uri.startsWith('ph://')) {
-    const {width = 1000, height = 1000} = selectedImage?.node.image ?? {};
-    const manipulatedPath = await PhotoManipulator.crop(uri, {
-      x: 0,
-      y: 0,
-      width,
-      height,
-    });
-    return manipulatedPath;
-  }
-  return uri;
-}
-
-async function loadSkiaImage(uri: string): Promise<SkImage | null> {
-  try {
-    const response = await fetch(uri);
-    const buffer = await response.arrayBuffer();
-    const skData = Skia.Data.fromBytes(new Uint8Array(buffer));
-    const skImage = Skia.Image.MakeImageFromEncoded(skData);
-    return skImage ?? null;
-  } catch (err) {
-    console.warn('Skia image load error', err);
-    return null;
-  }
-}
+import {useImageState} from '../../service/hooks/image-state.hook';
+import {useFilterOperations} from '../../service/hooks/filter-operations.hook';
 
 const GalleryDetailFilterPage = (): JSX.Element => {
   const canvasRef = useCanvasRef();
-
-  const [skiaImage, setSkiaImage] = useState<SkImage | null>(null);
-  const [imageSize, setImageSize] = useState({
-    width: displaySize,
-    height: displaySize,
-  });
-  const [activeFilter, setActiveFilter] = useState<FilterType>('original');
-  const [filterAmount, setFilterAmount] = useState(1);
-
-  const [galleryIndex, setGalleryIndex] = useRecoilState(
-    selectedGalleryIndexState,
-  );
-  const [editGalleryItems, setEditGalleryItems] = useRecoilState(
-    editedGalleryItemsState,
-  );
-  const [selectedImage, setSelectedImage] = useState<
-    PhotoIdentifier | undefined
-  >(editGalleryItems[galleryIndex]);
-
   const navigation = useNavigation<BasicNavigationProps>();
+  const {skiaImage, imageSize, selectedImage, displaySize} = useImageState();
+  const {
+    activeFilter,
+    filterAmount,
+    setFilterAmount,
+    isSliderNeeded,
+    currentSliderConfig,
+    applyFilter,
+    saveFilteredImage,
+  } = useFilterOperations();
 
-  const isSliderNeeded = FILTER_SETTINGS[activeFilter] !== undefined;
-  const currentSliderConfig = FILTER_SETTINGS[activeFilter];
+  const handleSaveFilteredImage = useCallback(() => {
+    saveFilteredImage(canvasRef, skiaImage, selectedImage);
+  }, [saveFilteredImage, canvasRef, skiaImage, selectedImage]);
 
-  const applyFilter = (filterName: FilterType) => {
-    if (!selectedImage) {
-      CustomAlert.simpleAlert('먼저 사진을 선택해주세요.');
-      return;
-    }
-    setActiveFilter(filterName);
+  const handleApplyFilter = useCallback(
+    (filterName: FilterType) => {
+      applyFilter(filterName, selectedImage);
+    },
+    [applyFilter, selectedImage],
+  );
 
-    const config = FILTER_SETTINGS[filterName];
-    setFilterAmount(config ? config.initial : 1);
-  };
-
-  const saveFilteredImage = useCallback(async () => {
-    if (!canvasRef.current || !skiaImage) {
-      CustomAlert.simpleAlert('저장할 사진이 없습니다.');
-      return;
-    }
-
-    try {
-      const snapshot = canvasRef.current.makeImageSnapshot();
-      const bytes = snapshot.encodeToBytes();
-      const fileName = `filtered_${Date.now()}.png`;
-      const filePath = `${RNFS.CachesDirectoryPath}/${fileName}`;
-
-      await writeFile(filePath, encodeBase64(bytes.buffer), 'base64');
-
-      const newImageObject = {
-        ...selectedImage!,
-        node: {
-          ...selectedImage!.node,
-          image: {
-            ...selectedImage!.node.image,
-            uri: 'file://' + filePath,
-            width: skiaImage.width(),
-            height: skiaImage.height(),
-            size: bytes.length,
-            filename: fileName,
-          },
-        },
-      };
-
-      const updatedGallery = editGalleryItems.map((e, idx) =>
-        idx === galleryIndex ? newImageObject : e,
-      );
-      setEditGalleryItems(updatedGallery);
-      navigation.goBack();
-    } catch (err) {
-      console.error('필터 저장 오류:', err);
-    }
-  }, [
-    canvasRef,
-    skiaImage,
-    selectedImage,
-    editGalleryItems,
-    galleryIndex,
-    setEditGalleryItems,
-    navigation,
-  ]);
-
-  useEffect(() => {
-    if (editGalleryItems && editGalleryItems.length > galleryIndex) {
-      setSelectedImage(editGalleryItems[galleryIndex]);
-    } else {
-      setSelectedImage(undefined);
-      CustomAlert.simpleAlert(
-        '이미지를 불러올 수 없습니다. 갤러리 항목을 확인해주세요.',
-      );
-    }
-  }, [editGalleryItems, galleryIndex]);
-
-  useEffect(() => {
-    const fetchImageSize = async () => {
-      if (selectedImage?.node.image.uri) {
-        let width = selectedImage.node.image.width;
-        let height = selectedImage.node.image.height;
-
-        if (!width || !height) {
-          try {
-            const size = await getImageSizeAsync(selectedImage.node.image.uri);
-            width = size.width;
-            height = size.height;
-          } catch (e) {
-            setImageSize({width: displaySize, height: displaySize});
-            return;
-          }
+  const HeaderComponent = useMemo(
+    () => (
+      <TopBar
+        title={'사진 편집'}
+        right={
+          <WritingHeaderRight
+            text={'완료'}
+            customAction={handleSaveFilteredImage}
+            disable={!selectedImage || activeFilter === 'original'}
+          />
         }
-
-        const aspectRatio = width / height;
-        setImageSize({width: displaySize, height: displaySize / aspectRatio});
-      } else {
-        setImageSize({width: displaySize, height: displaySize});
-      }
-    };
-
-    fetchImageSize();
-  }, [selectedImage]);
-
-  useEffect(() => {
-    (async () => {
-      if (selectedImage?.node.image.uri) {
-        const path = await copyContentUriToFile(selectedImage);
-        const img = await loadSkiaImage(path);
-        setSkiaImage(img);
-      }
-    })();
-  }, [selectedImage]);
+      />
+    ),
+    [handleSaveFilteredImage, selectedImage, activeFilter],
+  );
 
   useEffect(() => {
     navigation.setOptions({
-      header: () => (
-        <TopBar
-          title={'사진 편집'}
-          right={
-            <WritingHeaderRight
-              text={'완료'}
-              customAction={saveFilteredImage}
-              disable={!selectedImage || activeFilter === 'original'}
-            />
-          }
-        />
-      ),
+      header: () => HeaderComponent,
     });
-  }, [navigation, saveFilteredImage, selectedImage, activeFilter]);
+  }, [navigation, HeaderComponent]);
 
   return (
     <LoadingContainer isLoading={false}>
@@ -292,7 +124,7 @@ const GalleryDetailFilterPage = (): JSX.Element => {
           )}
         </ContentContainer>
 
-        <ContentContainer minHeight={40}>
+        <ContentContainer minHeight="40">
           {isSliderNeeded && currentSliderConfig && (
             <Slider
               minimumValue={currentSliderConfig.min}
@@ -326,7 +158,7 @@ const GalleryDetailFilterPage = (): JSX.Element => {
                     borderRadius: 5,
                   },
                 ]}
-                onPress={() => applyFilter(filterName)}>
+                onPress={() => handleApplyFilter(filterName as FilterType)}>
                 <ContentContainer borderRadius={5} alignCenter flex={1}>
                   <Title>{FILTER_LABELS[filterName as FilterType]}</Title>
                 </ContentContainer>
